Automations and bot/botta-master/requirements.txt
pandas
python-dotenv